Move ServerStatus to util package to remove tangle
