+++ neuro_san/service/http/handlers/health_check_handler.py
@@ -22,8 +22,8 @@ from importlib.metadata import version as library_version
 from importlib.metadata import PackageNotFoundError
 
 from tornado.web import RequestHandler
-from neuro_san.service.main_loop.server_status import ServerStatus
 from neuro_san.service.http.logging.http_logger import HttpLogger
+from neuro_san.service.util.server_status import ServerStatus
 
 
 class HealthCheckHandler(RequestHandler):
