+++ neuro_san/service/main_loop/server_main_loop.py
@@ -36,8 +36,8 @@ from neuro_san.service.grpc.grpc_agent_server import DEFAULT_REQUEST_LIMIT
 from neuro_san.service.grpc.grpc_agent_server import GrpcAgentServer
 from neuro_san.service.grpc.grpc_agent_service import GrpcAgentService
 from neuro_san.service.http.server.http_server import HttpServer
-from neuro_san.service.main_loop.server_status import ServerStatus
 from neuro_san.service.watcher.main_loop.storage_watcher import StorageWatcher
+from neuro_san.service.util.server_status import ServerStatus
 
 
 # pylint: disable=too-many-instance-attributes
